feat: 식당 관련 이미지 처리 구현

+++ src/main/java/com/example/favoriteschoolmeal/domain/restaurant/service/RestaurantService.java
@@ -4,14 +4,16 @@ import com.example.favoriteschoolmeal.domain.restaurant.controller.dto.CreateRes
 import com.example.favoriteschoolmeal.domain.restaurant.controller.dto.RestaurantResponse;
 import com.example.favoriteschoolmeal.domain.restaurant.domain.Restaurant;
 import com.example.favoriteschoolmeal.domain.restaurant.exeption.RestaurantExceptionType;
-import com.example.favoriteschoolmeal.domain.restaurant.exeption.RestaurantNotFoundException;
+import com.example.favoriteschoolmeal.domain.restaurant.exeption.RestaurantException;
 import com.example.favoriteschoolmeal.domain.restaurant.repository.RestaurantRepository;
+import com.example.favoriteschoolmeal.global.security.util.SecurityUtils;
 import jakarta.transaction.Transactional;
+
 import java.util.List;
 import java.util.Optional;
+
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
 
 @Service
 @RequiredArgsConstructor
@@ -20,9 +22,8 @@ public class RestaurantService {
 
     private final RestaurantRepository restaurantRepository;
 
-    public RestaurantResponse addRestaurant(CreateRestaurantRequest request,
-            MultipartFile thumbnail, MultipartFile menuImage) {
-        //TODO: file처리 추가
+    public RestaurantResponse addRestaurant(final CreateRestaurantRequest request) {
+        verifyRoleAdmin();
 
         Restaurant restaurant = Restaurant.builder()
                 .name(request.name())
@@ -30,22 +31,17 @@ public class RestaurantService {
                 .isOnCampus(request.isOnCampus())
                 .category(request.category())
                 .businessHours(request.businessHours())
-                //TODO: thumbnail, menuImage 처리 추가
-                .thumbnailId(1L) //임시로 1로 설정
+                .thumbnailUrl(request.thumbnailUrl())
+                .menuImageUrl(request.menuImageUrl())
                 .build();
 
         restaurantRepository.save(restaurant);
         return RestaurantResponse.of(restaurant);
     }
 
-    // TODO: PR에서 리팩토링 제안 예정 (주석 작성자: 조예림)
-    public RestaurantResponse findRestaurant(Long restaurantId) {
-        Restaurant restaurant = restaurantRepository.findById(restaurantId)
-                .orElseThrow(() -> new RestaurantNotFoundException(
-                        RestaurantExceptionType.RESTAURANT_NOT_FOUND));
-
+    public RestaurantResponse findRestaurant(final Long restaurantId) {
+        Restaurant restaurant = getRestaurantOrThrow(restaurantId);
         return RestaurantResponse.of(restaurant);
-
     }
 
     public List<RestaurantResponse> findAllRestaurant() {
@@ -53,31 +49,34 @@ public class RestaurantService {
         return restaurants.stream().map(RestaurantResponse::of).toList();
     }
 
-    public RestaurantResponse modifyRestaurant(Long restaurantId, CreateRestaurantRequest request,
-            MultipartFile thumbnail, MultipartFile menuImage) {
-        Restaurant restaurant = restaurantRepository.findById(restaurantId)
-                .orElseThrow(() -> new RestaurantNotFoundException(
-                        RestaurantExceptionType.RESTAURANT_NOT_FOUND));
-
-        //TODO: file처리 추가
-        Long thumbnailId = 1L; //임시로 1로 설정
-        Long menuImageId = 1L; //임시로 1로 설정
+    public RestaurantResponse modifyRestaurant(final Long restaurantId, final CreateRestaurantRequest request) {
+        verifyRoleAdmin();
+        Restaurant restaurant = getRestaurantOrThrow(restaurantId);
         restaurant.update(request.isOnCampus(), request.location(), request.category(),
-                request.name(), request.businessHours(), thumbnailId, menuImageId);
+                request.name(), request.businessHours(), request.thumbnailUrl(), request.menuImageUrl());
         return RestaurantResponse.of(restaurant);
     }
 
 
-    public Long deleteRestaurant(Long restaurantId) {
-        Restaurant restaurant = restaurantRepository.findById(restaurantId)
-                .orElseThrow(() -> new RestaurantNotFoundException(
-                        RestaurantExceptionType.RESTAURANT_NOT_FOUND));
-
+    public Long deleteRestaurant(final Long restaurantId) {
+        verifyRoleAdmin();
+        Restaurant restaurant = getRestaurantOrThrow(restaurantId);
         restaurantRepository.delete(restaurant);
         return restaurantId;
     }
 
-    public Optional<Restaurant> findRestaurantOptionally(Long restaurantId) {
+
+    public Optional<Restaurant> findRestaurantOptionally(final Long restaurantId) {
         return restaurantRepository.findById(restaurantId);
+    }
+
+    private Restaurant getRestaurantOrThrow(final Long restaurantId) {
+        return restaurantRepository.findById(restaurantId)
+                .orElseThrow(() -> new RestaurantException(
+                        RestaurantExceptionType.RESTAURANT_NOT_FOUND));
+    }
+
+    private void verifyRoleAdmin() {
+        SecurityUtils.checkAdminOrThrow(() -> new RestaurantException(RestaurantExceptionType.UNAUTHORIZED_ACCESS));
     }
 }
